extracting mockNumberConversionServiceObject to unit-test-utils
