+++ core/templates/components/forms/schema-based-editors/integration-tests/schema-based-editors.integration.spec.ts
@@ -41,6 +41,7 @@ import {DirectivesModule} from 'directives/directives.module';
 import {SharedPipesModule} from 'filters/shared-pipes.module';
 import {MaterialModule} from 'modules/material.module';
 import {DictSchema, UnicodeSchema} from 'services/schema-default-value.service';
+import {mockNumberConversionServiceObject} from 'tests/unit-test-utils';
 import {MockTranslateModule} from 'tests/unit-test-utils';
 import {NumberConversionService} from 'services/number-conversion.service';
 import {SchemaBasedBoolEditorComponent} from '../schema-based-bool-editor.component';
@@ -109,16 +110,7 @@ describe('Schema based editor', () => {
         {provide: TranslateService, useClass: MockTranslateService},
         {
           provide: NumberConversionService,
-          useValue: {
-            currentDecimalSeparator: () => '.',
-            convertToEnglishDecimal: (value: string | number | null) => {
-              if (value === null || value === undefined) {
-                return null;
-              }
-              const result = parseFloat(value.toString().replace(',', '.'));
-              return isNaN(result) ? null : result;
-            },
-          },
+          useValue: mockNumberConversionServiceObject,
         },
       ],
     }).compileComponents();
